website/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
asyncpg==0.30.0
python-multipart==0.0.12
